Enum anvil_core::eth::EthRequest

pub enum EthRequest {

    Web3ClientVersion(()),
    Web3Sha3(Bytes),
    EthChainId(()),
    EthNetworkId(()),
    NetListening(()),
    EthGasPrice(()),
    EthMaxPriorityFeePerGas(()),
    EthBlobBaseFee(()),
    EthAccounts(()),
    EthBlockNumber(()),
    EthGetBalance(Address, Option<BlockId>),
    EthGetStorageAt(Address, U256, Option<BlockId>),
    EthGetBlockByHash(B256, bool),
    EthGetBlockByNumber(BlockNumberOrTag, bool),
    EthGetTransactionCount(Address, Option<BlockId>),
    EthGetTransactionCountByHash(B256),
    EthGetTransactionCountByNumber(BlockNumberOrTag),
    EthGetUnclesCountByHash(B256),
    EthGetUnclesCountByNumber(BlockNumberOrTag),
    EthGetCodeAt(Address, Option<BlockId>),
    EthGetProof(Address, Vec<B256>, Option<BlockId>),
    EthSign(Address, Bytes),
    EthSignTransaction(Box<WithOtherFields<TransactionRequest>>),
    EthSignTypedData(Address, Value),
    EthSignTypedDataV3(Address, Value),
    EthSignTypedDataV4(Address, TypedData),
    EthSendTransaction(Box<WithOtherFields<TransactionRequest>>),
    EthSendRawTransaction(Bytes),
    EthCall(WithOtherFields<TransactionRequest>, Option<BlockId>, Option<StateOverride>),
    EthCreateAccessList(WithOtherFields<TransactionRequest>, Option<BlockId>),
    EthEstimateGas(WithOtherFields<TransactionRequest>, Option<BlockId>, Option<StateOverride>),
    EthGetTransactionByHash(TxHash),
    EthGetTransactionByBlockHashAndIndex(TxHash, Index),
    EthGetTransactionByBlockNumberAndIndex(BlockNumberOrTag, Index),
    EthGetRawTransactionByHash(TxHash),
    EthGetRawTransactionByBlockHashAndIndex(TxHash, Index),
    EthGetRawTransactionByBlockNumberAndIndex(BlockNumberOrTag, Index),
    EthGetTransactionReceipt(B256),
    EthGetBlockReceipts(BlockNumberOrTag),
    EthGetUncleByBlockHashAndIndex(B256, Index),
    EthGetUncleByBlockNumberAndIndex(BlockNumberOrTag, Index),
    EthGetLogs(Filter),
    EthNewFilter(Filter),
    EthGetFilterChanges(String),
    EthNewBlockFilter(()),
    EthNewPendingTransactionFilter(()),
    EthGetFilterLogs(String),
    EthUninstallFilter(String),
    EthGetWork(()),
    EthSubmitWork(B64, B256, B256),
    EthSubmitHashRate(U256, B256),
    EthFeeHistory(U256, BlockNumberOrTag, Vec<f64>),
    EthSyncing(()),
    DebugGetRawTransaction(TxHash),
    DebugTraceTransaction(B256, GethDebugTracingOptions),
    DebugTraceCall(WithOtherFields<TransactionRequest>, Option<BlockId>, GethDefaultTracingOptions),
    TraceTransaction(B256),
    TraceBlock(BlockNumberOrTag),
    ImpersonateAccount(Address),
    StopImpersonatingAccount(Address),
    AutoImpersonateAccount(bool),
    GetAutoMine(()),
    Mine(Option<U256>, Option<U256>),
    SetAutomine(bool),
    SetIntervalMining(u64),
    DropTransaction(B256),
    DropAllTransactions(),
    Reset(Option<Params<Option<Forking>>>),
    SetRpcUrl(String),
    SetBalance(Address, U256),
    SetCode(Address, Bytes),
    SetNonce(Address, U256),
    SetStorageAt(Address, U256, B256),
    SetCoinbase(Address),
    SetChainId(u64),
    SetLogging(bool),
    SetMinGasPrice(U256),
    SetNextBlockBaseFeePerGas(U256),
    EvmSetTime(U256),
    DumpState(()),
    LoadState(Bytes),
    NodeInfo(()),
    AnvilMetadata(()),
    EvmSnapshot(()),
    EvmRevert(U256),
    EvmIncreaseTime(U256),
    EvmSetNextBlockTimeStamp(U256),
    EvmSetBlockGasLimit(U256),
    EvmSetBlockTimeStampInterval(u64),
    EvmRemoveBlockTimeStampInterval(()),
    EvmMine(Option<Params<Option<MineOptions>>>),
    EvmMineDetailed(Option<Params<Option<MineOptions>>>),
    EthSendUnsignedTransaction(Box<WithOtherFields<TransactionRequest>>),
    EnableTraces(()),
    TxPoolStatus(()),
    TxPoolInspect(()),
    TxPoolContent(()),
    ErigonGetHeaderByNumber(BlockNumberOrTag),
    OtsGetApiLevel(()),
    OtsGetInternalOperations(B256),
    OtsHasCode(Address, BlockNumberOrTag),
    OtsTraceTransaction(B256),
    OtsGetTransactionError(B256),
    OtsGetBlockDetails(BlockNumberOrTag),
    OtsGetBlockDetailsByHash(B256),
    OtsGetBlockTransactions(u64, usize, usize),
    OtsSearchTransactionsBefore(Address, u64, usize),
    OtsSearchTransactionsAfter(Address, u64, usize),
    OtsGetTransactionBySenderAndNonce(Address, U256),
    OtsGetContractCreator(Address),
    RemovePoolTransactions(Address),

}

Represents ethereum JSON-RPC API

Variants

Web3ClientVersion(())

Web3Sha3(Bytes)

EthChainId(())

EthNetworkId(())

NetListening(())

EthGasPrice(())

EthMaxPriorityFeePerGas(())

EthBlobBaseFee(())

EthAccounts(())

EthBlockNumber(())

EthGetBalance(Address, Option<BlockId>)

EthGetStorageAt(Address, U256, Option<BlockId>)

EthGetBlockByHash(B256, bool)

EthGetBlockByNumber(BlockNumberOrTag, bool)

EthGetTransactionCount(Address, Option<BlockId>)

EthGetTransactionCountByHash(B256)

EthGetTransactionCountByNumber(BlockNumberOrTag)

EthGetUnclesCountByHash(B256)

EthGetUnclesCountByNumber(BlockNumberOrTag)

EthGetCodeAt(Address, Option<BlockId>)

EthGetProof(Address, Vec<B256>, Option<BlockId>)

Returns the account and storage values of the specified account including the Merkle-proof. This call can be used to verify that the data you are pulling from is not tampered with.

EthSign(Address, Bytes)

The sign method calculates an Ethereum specific signature with:

EthSignTransaction(Box<WithOtherFields<TransactionRequest>>)

EthSignTypedData(Address, Value)

Signs data via EIP-712.

EthSignTypedDataV3(Address, Value)

Signs data via EIP-712.

EthSignTypedDataV4(Address, TypedData)

Signs data via EIP-712, and includes full support of arrays and recursive data structures.

EthSendTransaction(Box<WithOtherFields<TransactionRequest>>)

EthSendRawTransaction(Bytes)

EthCall(WithOtherFields<TransactionRequest>, Option<BlockId>, Option<StateOverride>)

EthCreateAccessList(WithOtherFields<TransactionRequest>, Option<BlockId>)

EthEstimateGas(WithOtherFields<TransactionRequest>, Option<BlockId>, Option<StateOverride>)

EthGetTransactionByHash(TxHash)

EthGetTransactionByBlockHashAndIndex(TxHash, Index)

EthGetTransactionByBlockNumberAndIndex(BlockNumberOrTag, Index)

EthGetRawTransactionByHash(TxHash)

EthGetRawTransactionByBlockHashAndIndex(TxHash, Index)

EthGetRawTransactionByBlockNumberAndIndex(BlockNumberOrTag, Index)

EthGetTransactionReceipt(B256)

EthGetBlockReceipts(BlockNumberOrTag)

EthGetUncleByBlockHashAndIndex(B256, Index)

EthGetUncleByBlockNumberAndIndex(BlockNumberOrTag, Index)

EthGetLogs(Filter)

EthNewFilter(Filter)

Creates a filter object, based on filter options, to notify when the state changes (logs).

EthGetFilterChanges(String)

Polling method for a filter, which returns an array of logs which occurred since last poll.

EthNewBlockFilter(())

Creates a filter in the node, to notify when a new block arrives. To check if the state has changed, call eth_getFilterChanges.

EthNewPendingTransactionFilter(())

Creates a filter in the node, to notify when new pending transactions arrive. To check if the state has changed, call eth_getFilterChanges.

EthGetFilterLogs(String)

Returns an array of all logs matching filter with given id.

EthUninstallFilter(String)

Removes the filter, returns true if the filter was installed

EthGetWork(())

EthSubmitWork(B64, B256, B256)

EthSubmitHashRate(U256, B256)

EthFeeHistory(U256, BlockNumberOrTag, Vec<f64>)

EthSyncing(())

DebugGetRawTransaction(TxHash)

geth’s debug_getRawTransaction endpoint

DebugTraceTransaction(B256, GethDebugTracingOptions)

geth’s debug_traceTransaction endpoint

DebugTraceCall(WithOtherFields<TransactionRequest>, Option<BlockId>, GethDefaultTracingOptions)

geth’s debug_traceCall endpoint

TraceTransaction(B256)

Trace transaction endpoint for parity’s trace_transaction

TraceBlock(BlockNumberOrTag)

Trace transaction endpoint for parity’s trace_block

ImpersonateAccount(Address)

send transactions impersonating specific account and contract addresses.

StopImpersonatingAccount(Address)

Stops impersonating an account if previously set with anvil_impersonateAccount

AutoImpersonateAccount(bool)

Will make every account impersonated

GetAutoMine(())

Returns true if automatic mining is enabled, and false.

Mine(Option<U256>, Option<U256>)

Mines a series of blocks

Tuple Fields

0: Option<U256>

Number of blocks to mine, if not set 1 block is mined

1: Option<U256>

The time interval between each block in seconds, defaults to 1 seconds The interval is applied only to blocks mined in the given method invocation, not to blocks mined afterwards. Set this to 0 to instantly mine all blocks

SetAutomine(bool)

Enables or disables, based on the single boolean argument, the automatic mining of new blocks with each new transaction submitted to the network.

SetIntervalMining(u64)

Sets the mining behavior to interval with the given interval (seconds)

DropTransaction(B256)

Removes transactions from the pool

DropAllTransactions()

Removes transactions from the pool

Reset(Option<Params<Option<Forking>>>)

Reset the fork to a fresh forked state, and optionally update the fork config

SetRpcUrl(String)

Sets the backend rpc url

SetBalance(Address, U256)

Modifies the balance of an account.

SetCode(Address, Bytes)

Sets the code of a contract

SetNonce(Address, U256)

Sets the nonce of an address

SetStorageAt(Address, U256, B256)

Writes a single slot of the account’s storage

Tuple Fields

0: Address

1: U256

slot

2: B256

value

SetCoinbase(Address)

Sets the coinbase address

SetChainId(u64)

Sets the chain id

SetLogging(bool)

Enable or disable logging

SetMinGasPrice(U256)

Set the minimum gas price for the node

SetNextBlockBaseFeePerGas(U256)

Sets the base fee of the next block

EvmSetTime(U256)

Sets the specific timestamp Accepts timestamp (Unix epoch) with millisecond precision and returns the number of seconds between the given timestamp and the current time.

DumpState(())

Serializes the current state (including contracts code, contract’s storage, accounts properties, etc.) into a savable data blob

LoadState(Bytes)

Adds state previously dumped with DumpState to the current chain

NodeInfo(())

Retrieves the Anvil node configuration params

AnvilMetadata(())

Retrieves the Anvil node metadata.

EvmSnapshot(())

Snapshot the state of the blockchain at the current block.

Ref https://github.com/trufflesuite/ganache/blob/ef1858d5d6f27e4baeb75cccd57fb3dc77a45ae8/src/chains/ethereum/ethereum/RPC-METHODS.md#evm_snapshot

EvmRevert(U256)

Revert the state of the blockchain to a previous snapshot. Takes a single parameter, which is the snapshot id to revert to.

Ref https://github.com/trufflesuite/ganache/blob/ef1858d5d6f27e4baeb75cccd57fb3dc77a45ae8/src/chains/ethereum/ethereum/RPC-METHODS.md#evm_revert

EvmIncreaseTime(U256)

Jump forward in time by the given amount of time, in seconds.

EvmSetNextBlockTimeStamp(U256)

Similar to evm_increaseTime but takes the exact timestamp that you want in the next block

EvmSetBlockGasLimit(U256)

Set the exact gas limit that you want in the next block

EvmSetBlockTimeStampInterval(u64)

Similar to evm_increaseTime but takes sets a block timestamp interval.

The timestamp of the next block will be computed as lastBlock_timestamp + interval.

EvmRemoveBlockTimeStampInterval(())

Removes a anvil_setBlockTimestampInterval if it exists

EvmMine(Option<Params<Option<MineOptions>>>)

Mine a single block

EvmMineDetailed(Option<Params<Option<MineOptions>>>)

Mine a single block and return detailed data

This behaves exactly as EvmMine but returns different output, for compatibility reasons this is a separate call since evm_mine is not an anvil original.

EthSendUnsignedTransaction(Box<WithOtherFields<TransactionRequest>>)

Execute a transaction regardless of signature status

EnableTraces(())

Turn on call traces for transactions that are returned to the user when they execute a transaction (instead of just txhash/receipt)

TxPoolStatus(())

Returns the number of transactions currently pending for inclusion in the next block(s), as well as the ones that are being scheduled for future execution only. Ref: https://geth.ethereum.org/docs/rpc/ns-txpool#txpool_status

TxPoolInspect(())

Returns a summary of all the transactions currently pending for inclusion in the next block(s), as well as the ones that are being scheduled for future execution only. Ref: https://geth.ethereum.org/docs/rpc/ns-txpool#txpool_inspect

TxPoolContent(())

Returns the details of all transactions currently pending for inclusion in the next block(s), as well as the ones that are being scheduled for future execution only. Ref: https://geth.ethereum.org/docs/rpc/ns-txpool#txpool_content

ErigonGetHeaderByNumber(BlockNumberOrTag)

Otterscan’s ots_getApiLevel endpoint Otterscan currently requires this endpoint, even though it’s not part of the ots_* https://github.com/otterscan/otterscan/blob/071d8c55202badf01804f6f8d53ef9311d4a9e47/src/useProvider.ts#L71 Related upstream issue: https://github.com/otterscan/otterscan/issues/1081

OtsGetApiLevel(())

Otterscan’s ots_getApiLevel endpoint Used as a simple API versioning scheme for the ots_* namespace

OtsGetInternalOperations(B256)

Otterscan’s ots_getInternalOperations endpoint Traces internal ETH transfers, contracts creation (CREATE/CREATE2) and self-destructs for a certain transaction.

OtsHasCode(Address, BlockNumberOrTag)

Otterscan’s ots_hasCode endpoint Check if an ETH address contains code at a certain block number.

OtsTraceTransaction(B256)

Otterscan’s ots_traceTransaction endpoint Trace a transaction and generate a trace call tree.

OtsGetTransactionError(B256)

Otterscan’s ots_getTransactionError endpoint Given a transaction hash, returns its raw revert reason.

OtsGetBlockDetails(BlockNumberOrTag)

Otterscan’s ots_getBlockDetails endpoint Given a block number, return its data. Similar to the standard eth_getBlockByNumber/Hash method, but can be optimized by excluding unnecessary data such as transactions and logBloom

OtsGetBlockDetailsByHash(B256)

Otterscan’s ots_getBlockDetails endpoint Same as ots_getBlockDetails, but receiving a block hash instead of number

OtsGetBlockTransactions(u64, usize, usize)

Otterscan’s ots_getBlockTransactions endpoint Gets paginated transaction data for a certain block. Return data is similar to eth_getBlockBy* + eth_getTransactionReceipt.

OtsSearchTransactionsBefore(Address, u64, usize)

Otterscan’s ots_searchTransactionsBefore endpoint Address history navigation. searches backwards from certain point in time.

OtsSearchTransactionsAfter(Address, u64, usize)

Otterscan’s ots_searchTransactionsAfter endpoint Address history navigation. searches forward from certain point in time.

OtsGetTransactionBySenderAndNonce(Address, U256)

Otterscan’s ots_getTransactionBySenderAndNonce endpoint Given a sender address and a nonce, returns the tx hash or null if not found. It returns only the tx hash on success, you can use the standard eth_getTransactionByHash after that to get the full transaction data.

OtsGetContractCreator(Address)

Otterscan’s ots_getTransactionBySenderAndNonce endpoint Given an ETH contract address, returns the tx hash and the direct address who created the contract.

RemovePoolTransactions(Address)

Removes transactions from the pool by sender origin.