anvil_core::eth::transaction

Type Alias ReceiptResponse

pub type ReceiptResponse = TransactionReceipt<TypedReceipt<Receipt<Log>>>;

Aliased Type

pub struct ReceiptResponse {
    pub inner: TypedReceipt<Receipt<Log>>,
    pub transaction_hash: FixedBytes<32>,
    pub transaction_index: Option<u64>,
    pub block_hash: Option<FixedBytes<32>>,
    pub block_number: Option<u64>,
    pub gas_used: u64,
    pub effective_gas_price: u128,
    pub blob_gas_used: Option<u64>,
    pub blob_gas_price: Option<u128>,
    pub from: Address,
    pub to: Option<Address>,
    pub contract_address: Option<Address>,
}

Fields
inner: TypedReceipt<Receipt<Log>>

The receipt envelope, which contains the consensus receipt data.

transaction_hash: FixedBytes<32>

Transaction Hash.

transaction_index: Option<u64>

Index within the block.

block_hash: Option<FixedBytes<32>>

Hash of the block this transaction was included within.

block_number: Option<u64>

Number of the block this transaction was included within.

gas_used: u64

Gas used by this transaction alone.

effective_gas_price: u128

The price paid post-execution by the transaction (i.e. base fee + priority fee). Both fields in 1559-style transactions are maximums (max fee + max priority fee), the amount that’s actually paid by users can only be determined post-execution

blob_gas_used: Option<u64>

Blob gas used by the eip-4844 transaction

This is None for non eip-4844 transactions

blob_gas_price: Option<u128>

The price paid by the eip-4844 transaction per blob gas.

from: Address

Address of the sender

to: Option<Address>

Address of the receiver. None when its a contract creation transaction.

contract_address: Option<Address>

Contract address created, or None if not a deployment.

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 592 bytes