Module anvil_core::eth::trie

Utility functions for Ethereum adapted from https://github.dev/rust-blockchain/ethereum/blob/755dffaa4903fbec1269f50cde9863cf86269a14/src/util.rs

Modules

node

trie_constants

Constants specific to encoding with external value node support.

Macros

test_layouts

Apply a test method on every test layouts.

test_layouts_no_meta

Apply a test method on every test layouts.

Structs

AllowEmptyLayout

Trie that allows empty values.

Bitmap

Children bitmap codec for radix 16 trie.

ExtensionLayout

Trie layout using extension nodes.

GenericNoExtensionLayout

Trie layout without extension nodes, allowing generic hasher.

HashedValueNoExt

No extension trie with no hashed value.

HashedValueNoExtThreshold

No extension trie which stores value above a static size as external node.

KeccakHasher

Concrete Hasher impl for the Keccak-256 hash

ReferenceNodeCodec

Simple reference implementation of a NodeCodec.

ReferenceNodeCodecNoExt

Simple reference implementation of a NodeCodec. Even if implementation follows initial specification of https://github.com/w3f/polkadot-re-spec/issues/8, this may not follow it in the future, it is mainly the testing codec without extension node.

ReferenceNodeCodecNoExtMeta

ReferenceTrieStream

Reference implementation of a TrieStream with extension nodes.

ReferenceTrieStreamNoExt

Reference implementation of a TrieStream without extension.

Constants

KECCAK_NULL_RLP

The KECCAK of the RLP encoding of empty data.

Traits

TrieStream

Byte-stream oriented trait for constructing closed-form tries.

Functions

calc_root

Trie builder root calculation utility.

calc_root_build

Trie builder trie building utility.

compare_implementations

Compare trie builder and in memory trie.

compare_implementations_unordered

compare_implementations_no_extension for unordered input (trie_root does ordering before running when trie_build expect correct ordering).

compare_insert_remove

Testing utility that uses some periodic removal over its input test data.

compare_root

Compare trie builder and trie root implementations.

compare_unhashed

Compare trie builder and trie root unhashed implementations.

compare_unhashed_no_extension

Compare trie builder and trie root unhashed implementations. This uses the variant without extension nodes.

ordered_trie_root

Generates a trie root hash for a vector of values

reference_trie_root

reference_trie_root_iter_build

sec_trie_root

Generates a key-hashed (secure) trie root hash for a vector of key-value tuples.

trie_root

Generates a trie root hash for a vector of key-value tuples

Type Aliases

NoExtensionLayout

Trie layout without extension nodes.

RefFatDB

RefFatDBMut

RefHasher

Reference hasher is a keccak hasher.

RefLookup

RefLookupNoExt

RefSecTrieDB

RefSecTrieDBMut

RefTrieDB

RefTrieDBMut

RefTrieDBMutAllowEmpty

RefTrieDBMutNoExt